most of the new version
still need to modify the skeletons for student-supplied code
and the documentation

+++ src/spatial/trees/SpatialQuerySolver.java
@@ -2,7 +2,6 @@ package spatial.trees;
 import spatial.kdpoint.KDPoint;
 import spatial.knnutils.BoundedPriorityQueue;
 
-import java.math.BigDecimal;
 import java.util.Collection;
 
 /**
@@ -13,7 +12,7 @@ import java.util.Collection;
  *
  * <p><b>YOU SHOULD ***NOT*** EDIT THIS INTERFACE!</b> If you do, you risk <b>not passing our tests!</b></p>
  *
- * @author <a href="https://github.com/JasonFil">Jason Filippou</a>
+ * @author <a href="https://github.com/jasonfilippou">Jason Filippou</a>
  *
  * @see KDPoint
  * @see BoundedPriorityQueue
@@ -21,21 +20,21 @@ import java.util.Collection;
  */
 public interface SpatialQuerySolver {
 
-    /**
-     * Performs a range query. Returns all the {@link KDPoint}s whose {@link KDPoint#distanceSquared(KDPoint) distanceSquared} from
-     * p is at most range squared, <b>INCLUSIVE</b>.
+     /**
+     * Performs a range query. Returns all the {@link KDPoint}s whose {@link KDPoint#euclideanDistance(KDPoint) distanceSquared} from
+     * p is at most range, <b>INCLUSIVE</b>.
      * @param p The query {@link KDPoint}.
-     * @param range The maximum {@link KDPoint#distanceSquared(KDPoint, KDPoint) distance} from p
+     * @param range The maximum {@link KDPoint#euclideanDistance(KDPoint, KDPoint) distance} from p
      * that we allow a {@link KDPoint} to have if it should be part of the solution.
      * @return A {@link Collection} over all {@link KDPoint}s which satisfy our query. The
      * {@link Collection} will be empty if there are no points which satisfy the query.
      * @see KDPoint
      * @see Collection
      */
-    Collection<KDPoint> range(KDPoint p, BigDecimal range);
+    Collection<KDPoint> range(KDPoint p, double range);
 
     /** Performs a nearest neighbor query. Returns the {@link KDPoint} which is closest to
-     * p, as dictated by {@link KDPoint#distanceSquared(KDPoint) distanceSquared(KDPoint p)}.
+     * p, as dictated by {@link KDPoint#euclideanDistance(KDPoint) distanceSquared(KDPoint p)}.
      * @param p The query {@link KDPoint}.
      * @return The solution to the nearest neighbor query. This method will return null if
      * there are no points other than p in the tree.
@@ -46,7 +45,7 @@ public interface SpatialQuerySolver {
     /**
      * Performs a k-nearest neighbors query on the SpatialTree. Returns the <em>k</em>
      * {@link KDPoint}s which are nearest to p, as dictated by
-     * {@link KDPoint#distanceSquared(KDPoint)}.
+     * {@link KDPoint#euclideanDistance(KDPoint)}.
      * The {@link KDPoint}s are sorted in ascending order of distanceSquared.
      * @param k A positive integer denoting the amount of neighbors to return.
      * @param p The query point.
